update script structure and preparing for larger reconstructions
